requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.2.5
django-appconf==1.0.3
django-compressor==2.3
django-cors-headers==3.1.0
django-environ==0.4.5
django-sass-processor==0.7.3
django-sendgrid-v5==0.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
future==0.17.1
isort==4.3.21
lazy-object-proxy==1.4.2
libsass==0.19.2
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.3.1
python-http-client==3.2.1
pytz==2019.2
rcssmin==1.0.6
rjsmin==1.1.0
sendgrid==6.1.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
django-bootstrap4==1.0.1
